@@ -14,26 +14,24 @@
   "private": true,
   "dependencies": {
     "@agm/core": "latest",
-    "@angular/animations": "^7.0.2",
+    "@angular/animations": "^8.2.14",
     "@angular/cdk": "^7.0.2",
-    "@angular/common": "^7.0.2",
-    "@angular/compiler": "^7.0.2",
-    "@angular/core": "^7.0.2",
-    "@angular/fire": "^5.1.0",
+    "@angular/common": "^8.2.14",
+    "@angular/compiler": "^8.2.14",
+    "@angular/core": "^8.2.14",
+    "@angular/fire": "^5.4.2",
     "@angular/flex-layout": "^7.0.0-beta.19",
-    "@angular/forms": "^7.0.2",
-    "@angular/http": "^7.0.2",
+    "@angular/forms": "^8.2.14",
     "@angular/material": "^7.0.2",
     "@angular/material-moment-adapter": "^7.0.2",
-    "@angular/platform-browser": "^7.0.2",
-    "@angular/platform-browser-dynamic": "^7.0.2",
-    "@angular/router": "^7.0.2",
-    "@ngmodule/material-carousel": "^0.3.1",
-    "angular2-image-gallery": "^7.0.0",
+    "@angular/platform-browser": "^8.2.14",
+    "@angular/platform-browser-dynamic": "^8.2.14",
+    "@angular/router": "^8.2.14",
+    "@ngmodule/material-carousel": "0.6.0",
     "bootstrap": "3.3.7",
     "core-js": "^2.4.1",
-    "firebase": "^5.5.6",
-    "firebaseui": "3.0.0",
+    "firebase": "^5.11.1",
+    "firebaseui": "^4.1.0",
     "gm": "^1.23.1",
     "hammerjs": "2.0.8",
     "lodash": "^4.17.15",
@@ -41,21 +39,21 @@
     "moment": "latest",
     "ngx-bootstrap": "1.7.1",
     "onecolor": "^3.1.0",
-    "rxjs": "^6.3.3",
-    "rxjs-compat": "^6.1.0",
-    "zone.js": "^0.8.26"
+    "rxjs": "^6.5.4",
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.6.0",
+    "@angular-devkit/build-angular": "~0.803.25",
     "@angular-devkit/core": "latest",
-    "@angular/cli": "^7.0.3",
-    "@angular/compiler-cli": "^7.0.2",
-    "@angular/language-service": "^7.0.2",
+    "@angular/cli": "^8.3.25",
+    "@angular/compiler-cli": "^8.2.14",
+    "@angular/language-service": "^8.2.14",
     "@types/jasmine": "~2.8.3",
     "@types/jasminewd2": "~2.0.2",
     "@types/lodash": "^4.14.149",
     "@types/node": "~6.0.60",
-    "codelyzer": "^4.0.1",
+    "angular2-image-gallery": "^7.0.0",
+    "codelyzer": "^5.0.1",
     "jasmine-core": "~2.8.0",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "~2.0.0",
@@ -66,6 +64,6 @@
     "protractor": "~5.1.2",
     "ts-node": "~4.1.0",
     "tslint": "~5.9.1",
-    "typescript": "3.1.5"
+    "typescript": "3.5.3"
   }
 }
