@@ -1,13 +1,12 @@
-import {BrowserModule} from "@angular/platform-browser";
-import {NgModule} from "@angular/core";
-import {FormsModule} from "@angular/forms";
-import {HttpModule} from "@angular/http";
-import {BrowserAnimationsModule} from "@angular/platform-browser/animations";
-import {ImageService} from "./services/image.service";
-import {GalleryComponent} from "./gallery/gallery.component";
-import {ViewerComponent} from "./viewer/viewer.component";
-import {DemoComponent} from "./demo/demo.component";
-import {CommonModule} from "@angular/common";
+import {NgModule} from '@angular/core';
+import {FormsModule} from '@angular/forms';
+import {BrowserAnimationsModule} from '@angular/platform-browser/animations';
+import {ImageService} from './services/image.service';
+import {GalleryComponent} from './gallery/gallery.component';
+import {ViewerComponent} from './viewer/viewer.component';
+import {DemoComponent} from './demo/demo.component';
+import {CommonModule} from '@angular/common';
+import {HttpClientModule} from '@angular/common/http';
 
 
 @NgModule({
@@ -19,7 +18,7 @@ import {CommonModule} from "@angular/common";
   imports: [
     CommonModule,
     FormsModule,
-    HttpModule,
+    HttpClientModule,
     BrowserAnimationsModule
   ],
   providers: [
