@@ -1,7 +1,6 @@
-import { Injectable } from '@angular/core';
-import * as firebase from 'firebase';
+import {Injectable} from '@angular/core';
 import * as firebaseui from 'firebaseui';
-import { AngularFireAuth } from '@angular/fire/auth';
+import {AngularFireAuth} from '@angular/fire/auth';
 
 @Injectable()
 export class FirebaseAuthUiService {
